refactor backend to group sections together
