+++ busy_beaver/backend.py
@@ -1,15 +1,18 @@
+from datetime import datetime, timedelta
 import logging
 import os
 import time
+from typing import List
 from urllib.parse import urlencode
 import uuid
 
+from sqlalchemy import and_
+import pytz
 import requests
 
-from . import api, db
+from . import api, db, github_stats
 from .adapters.slack import SlackAdapter
 from .models import ApiUser, User
-from .post_summary_stats import post_summary
 
 logger = logging.getLogger(__name__)
 
@@ -227,11 +230,30 @@ class PublishGitHubSummaryResource:
         if "channel" not in data:
             logger.error("[Busy-Beaver] Post GitHub Summary Request -- need channel in JSON body")
             return
-        channel = data["channel"]
-        post_summary(channel=channel)
+        post_github_summary_to_slack(data["channel"])
 
         logger.info("[Busy-Beaver] Post GitHub Summary -- kicked-off")
         resp.media = {"run": "kicked_off"}
 
 
 api.add_route("/github-summary", PublishGitHubSummaryResource())
+
+
+@api.background.task
+def post_github_summary_to_slack(channel: str) -> None:
+    boundary_dt = utc_now_minus(timedelta(days=1))
+    channel_id = slack.get_channel_id(channel)
+    members = slack.get_channel_members(channel_id)
+
+    users: List[User] = db.query(User).filter(
+        and_(User.slack_id.in_(members), User.github_username.isnot(None))
+    ).all()
+    message = ""
+    for user in users:
+        message += github_stats.generate_summary(user, boundary_dt)
+
+    slack.post_message(channel_id, message)
+
+
+def utc_now_minus(period: timedelta):
+    return pytz.utc.localize(datetime.utcnow()) - period
